:lipstick: rename files to camel case and fix loader generation
