+++ personalization-webcomponents/index-checklist-detail.html
@@ -30,7 +30,7 @@
 
     <!-- built files will be auto injected -->
     <script
-      src="src/checklist-detail-webcomponent.ts"
+      src="src/ChecklistDetailWebcomponent.ts"
       type="module"
     ></script>
   </head>
