+++ personalization-webcomponents/index-checklist-overview.html
@@ -30,7 +30,7 @@
 
     <!-- built files will be auto injected -->
     <script
-      src="src/checklist-overview-webcomponent.ts"
+      src="src/ChecklistOverviewWebcomponent.ts"
       type="module"
     ></script>
   </head>
