+++ personalization-webcomponents/index-checklist-preview.html
@@ -30,7 +30,7 @@
 
     <!-- built files will be auto injected -->
     <script
-      src="src/checklist-preview-webcomponent.ts"
+      src="src/ChecklistPreviewWebcomponent.ts"
       type="module"
     ></script>
   </head>
