+++ personalization-webcomponents/index-my-checklists.html
@@ -30,7 +30,7 @@
 
     <!-- built files will be auto injected -->
     <script
-      src="src/my-checklists-webcomponent.ts"
+      src="src/MyChecklistsWebcomponent.ts"
       type="module"
     ></script>
   </head>
