+++ personalization-webcomponents/processes/post-build.js
@@ -27,7 +27,7 @@ import manifest from '../dist/src/.vite/manifest.json' with {type: 'json'};
 
 // Required filename content to be treated as webcomponent
 const REQUIRED_PREFIX = 'src/';
-const REQUIRED_SUFFIX = '-webcomponent.ts';
+const REQUIRED_SUFFIX = 'Webcomponent.ts';
 
 for (const key in manifest) {
     if (key.startsWith(REQUIRED_PREFIX) && key.endsWith(REQUIRED_SUFFIX)) {
