+++ personalization-webcomponents/src/ChecklistDetail.ce.vue
