+++ personalization-webcomponents/src/ChecklistOverview.ce.vue
