+++ personalization-webcomponents/src/ChecklistPreview.ce.vue
